requirements.txt
# Web framework and server
fastapi==0.109.2
uvicorn[standard]==0.27.1

# Database
motor==3.3.2
beanie==1.25.0
pymongo[srv]==4.6.1

# Task Queue
celery==5.3.6
redis==5.0.1

# Environment and HTTP
python-dotenv==1.0.1
httpx==0.26.0
backoff==2.2.1

# Utilities
pydantic==2.6.1
python-dateutil==2.8.2

# For type checking and development
types-python-dateutil==2.8.19.20240106
typing-extensions==4.9.0

aioredis==2.0.1

sse_starlette==2.1.3

pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0

# requests==2.32.3
# bs4==0.0.2